Fix build errors and ensure clean compilation
- VectorRenderer: SKColor.Alpha is read-only; use WithAlpha() instead
- GridRenderer: explicit float casts for double*float arithmetic in Zoom
- SkiaCanvasControl: remove unused WorldClick event
- Tests: add GlobalUsings.cs with Xunit import, fix GridRendererTests
  float/double BeApproximately type mismatches

Build: 0 errors, 0 warnings (both projects)
Tests require Windows runtime (Microsoft.WindowsDesktop.App)

+++ src/LinearAlgebra.App/Rendering/GridRenderer.cs
@@ -35,8 +35,8 @@ public class GridRenderer
         _pixelsPerUnit *= factor;
         _pixelsPerUnit = Math.Clamp(_pixelsPerUnit, 10f, 200f);
 
-        _originX = focusX - worldX * _pixelsPerUnit;
-        _originY = focusY + worldY * _pixelsPerUnit;
+        _originX = focusX - (float)(worldX * _pixelsPerUnit);
+        _originY = focusY + (float)(worldY * _pixelsPerUnit);
     }
 
     public float WorldToScreenX(double worldX) => _originX + (float)(worldX * _pixelsPerUnit);
